Add test file for utils.makeefi library
This moves the obsolete test from the test_coreos_live_artifacts_mono.py
test into a new test_util_makeefi.py test file.
